refactored viewsummary with new components for SelectWeek Form, RecentPLTable, and AddWeekForm
